+++ app/__init__.py
@@ -1,0 +1,14 @@
+from flask import Flask
+
+def create_app():
+    app = Flask(__name__)
+
+    from app.main import bp as main_bp
+    app.register_blueprint(main_bp)
+
+    @app.route('/api/status')
+    def status():
+        return {'status':'online'}, 200
+
+    return app
+
